Add support for multiple light sources, minor updates

+++ src/Graphics/OpenGL/VertexArray.cpp
@@ -16,7 +16,7 @@ void VertexArray::buffer_mesh(const BasicMesh& mesh)
     assert(id);
     GLBuffer vbo;
     GLBuffer ebo;
-    indices_ = mesh.indices.size();
+    indices_ = static_cast<GLint>(mesh.indices.size());
 
     // Element buffer
     glNamedBufferStorage(ebo.id, mesh.indices.size() * sizeof(GLuint), mesh.indices.data(),
